feat: add projection support to schema decoding benchmark
